feat(cart): implement shopping cart APIs with JWT user authentication and stock validation

+++ src/main/java/com/zone/agri/service/ProductService.java
@@ -216,7 +216,7 @@ public class ProductService {
         }
 
         // 2. Kiểm tra tồn kho
-        Long totalStock = inventoryRepository.sumQuantityByProductId(id);
+        Integer totalStock = inventoryRepository.sumQuantityByProductId(id);
         if (totalStock != null && totalStock > 0) {
             throw new ConflictException("Sản phẩm vẫn còn tồn kho, không thể xóa.");
         }
